#170 Provide a progress monitor - refactorings
