+++ drombler-fx-core-application/src/main/java/org/drombler/fx/core/application/impl/MainWindowOnCloseRequestHandler.java
@@ -1,13 +1,13 @@
 package org.drombler.fx.core.application.impl;
 
-import org.drombler.fx.core.application.WindowOnCloseRequestHandlerProvider;
 import javafx.stage.Stage;
-import org.apache.felix.scr.annotations.Activate;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.Deactivate;
-import org.apache.felix.scr.annotations.Reference;
 import org.drombler.acp.startup.main.MainWindowProvider;
+import org.drombler.fx.core.application.WindowOnCloseRequestHandlerProvider;
 import org.osgi.service.component.ComponentContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  *
@@ -21,22 +21,6 @@ public class MainWindowOnCloseRequestHandler {
 
     @Reference
     private WindowOnCloseRequestHandlerProvider windowOnCloseRequestHandlerProvider;
-
-    protected void bindMainWindowProvider(MainWindowProvider<Stage> mainWindowProvider) {
-        this.mainWindowProvider = mainWindowProvider;
-    }
-
-    protected void unbindMainWindowProvider(MainWindowProvider<Stage> mainWindowProvider) {
-        this.mainWindowProvider = null;
-    }
-
-    protected void bindWindowOnCloseRequestHandlerProvider(WindowOnCloseRequestHandlerProvider windowOnCloseRequestHandlerProvider) {
-        this.windowOnCloseRequestHandlerProvider = windowOnCloseRequestHandlerProvider;
-    }
-
-    protected void unbindWindowOnCloseRequestHandlerProvider(WindowOnCloseRequestHandlerProvider windowOnCloseRequestHandlerProvider) {
-        this.windowOnCloseRequestHandlerProvider = null;
-    }
 
     @Activate
     protected void activate(ComponentContext context) {
