+++ drombler-fx-core-standard-desktop-classic/src/main/java/org/drombler/fx/core/standard/desktop/classic/impl/ContentHandler.java
@@ -19,8 +19,8 @@ import org.apache.felix.scr.annotations.Activate;
 import org.apache.felix.scr.annotations.Component;
 import org.apache.felix.scr.annotations.Deactivate;
 import org.apache.felix.scr.annotations.Reference;
-import org.drombler.acp.startup.main.ApplicationExecutorProvider;
-import org.drombler.fx.startup.main.ApplicationContentProvider;
+import org.drombler.acp.core.commons.util.concurrent.ApplicationExecutorProvider;
+import org.drombler.fx.core.application.ApplicationContentProvider;
 import org.osgi.service.component.ComponentContext;
 
 /**
